+++ src/component/Game/GameError.ts
